# application/services/custom_services/testfeature/src/models/entitymodels/testentity.ts
export interface testentity 
{
   created_date: { type: Date },
   created_by: { type: String },
   last_modified_by: { type: String },
   last_modified_date: { type: Date },
   name: { type: String },
   corprate: { type: String },
   modules: { type: String }
}